Move file `data_type/follow_trajectory.[ch]pp` into directory `behavior`
