+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/follow_trajectory_action.hpp
@@ -21,7 +21,7 @@
 #include <openscenario_interpreter/syntax/trajectory_following_mode.hpp>
 #include <openscenario_interpreter/syntax/trajectory_ref.hpp>
 #include <pugixml.hpp>
-#include <traffic_simulator/data_type/follow_trajectory.hpp>
+#include <traffic_simulator/behavior/follow_trajectory.hpp>
 
 namespace openscenario_interpreter
 {
