+++ simulation/traffic_simulator/include/traffic_simulator/entity/entity_base.hpp
@@ -22,8 +22,8 @@
 #include <optional>
 #include <queue>
 #include <string>
+#include <traffic_simulator/behavior/follow_trajectory.hpp>
 #include <traffic_simulator/behavior/longitudinal_speed_planning.hpp>
-#include <traffic_simulator/data_type/follow_trajectory.hpp>
 #include <traffic_simulator/data_type/lane_change.hpp>
 #include <traffic_simulator/data_type/speed_change.hpp>
 #include <traffic_simulator/hdmap_utils/hdmap_utils.hpp>
