+++ simulation/traffic_simulator/src/behavior/follow_trajectory.cpp
@@ -15,7 +15,6 @@
 #include <quaternion_operation/quaternion_operation.h>
 
 #include <arithmetic/floating_point/comparison.hpp>
-#include <cmath>
 #include <geometry/vector3/hypot.hpp>
 #include <geometry/vector3/norm.hpp>
 #include <geometry/vector3/normalize.hpp>
@@ -23,7 +22,7 @@
 #include <geometry/vector3/truncate.hpp>
 #include <iostream>
 #include <scenario_simulator_exception/exception.hpp>
-#include <traffic_simulator/data_type/follow_trajectory.hpp>
+#include <traffic_simulator/behavior/follow_trajectory.hpp>
 
 namespace traffic_simulator
 {
